support for ref specs, fixed issue, hex-dump, set version to 1.0.2
Adding support for dynamic specs where the length is stored in
another spec, fixing a size issue with assoc spec, adding hex-dump
function, set version to 1.0.2.

Changes:
o changed the map containing the assoc spec so that it keeps
  insertion order even for assoc specs larger than 10
o added a new protocol ISpecWithRef which is an either-or
  with ISpec. Not sure about the best way to model this,
  input welcomed.
o added new code to AssociativeSpec and IndexedSpec to
  support the new ISpecWithRef type of spec where the
  contained spec needs to be aware of the containers
  data / structure.
o added new spec types ref-string* and ref-bytes* for
  dynamic types where the length is stored in another
  spec.
o added a hex-dump function capable of printing a human
  readable dump of nio ByteBuffers, netty ByteBufs, byte
  arrays, and Strings. Buffy has this. If accepted, octet
  will have one too (though faster, terser, and more
  configurable :). I find this very useful when trying
  to understand a binary format.
o added tests for the new ref spec types.

Not done: documentation